+++ lib/src/dart_formatter.dart
@@ -102,6 +102,7 @@ class DartFormatter {
 
     // Parse it.
     var parser = new Parser(stringSource, errorListener);
+    parser.parseGenericMethods = true;
 
     AstNode node;
     if (source.isCompilationUnit) {