+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add support for assert() in constructor initializers.
 * Correctly indent the right-hand side of `is` and `as` expressions.
 * Avoid splitting in index operators when possible.
+* Support generic methods (#556).
 
 # 0.2.11+1